added parameters

+++ albums_migration.sql
@@ -7,7 +7,7 @@ CREATE TABLE albums(
   artist VARCHAR(100) NOT NULL,
   name VARCHAR(50) NOT NULL,
   release_date INT UNSIGNED,
-  sales FLOAT,
+  sales FLOAT(6,2),
   genre VARCHAR(20),
   PRIMARY KEY(id)
  );
